resume_screener_project/resume_screener/requirements.txt
flask>=2.3.0
PyPDF2>=3.0.0
docx2txt>=0.8
python-docx>=0.8.11
spacy>=3.6.0
nltk>=3.8.1
scikit-learn>=1.3.0
numpy>=1.24.0
